# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: PERSONAL-3 Alpaca Trading System — Operator API
  version: 1.1.0
  license:
    name: Proprietary — PERSONAL-3 operator only
    identifier: LicenseRef-PERSONAL-3-Proprietary
  description: |
    Ганц operator-т зориулсан арилжааны системийн REST API.

    **Бүх хариунд хүчинтэй гурван үл хөдлөх дүрэм:**
    1. Мөнгөн утга бүр **тэмдэгт мөр** хэлбэрийн fixed-point decimal
       (`"10234.56"`). Float ХЭЗЭЭ Ч биш (AC-25).
    2. Timestamp бүр UTC, RFC 3339, `Z` төгсгөлтэй (AC-26).
    3. Өгөгдөлтэй хариу бүр `source` талбартай ба нэг хариу дотор
       ХОЛИМОГ `source` байхыг хориглоно (AC-20, AC-21).

    **Аюулгүйн гэрээ:** энэ API-д Alpaca руу order илгээх ЦОРЫН ГАНЦ зам нь
    `POST /orders/manual` ба approval queue-ийн `approve` хоёр. Хоёул дотооддоо
    Risk Agent → Execution модулиар дамжина. LLM нь энэ API-д огт хүрэхгүй —
    LLM зөвхөн Tool Contract (энэ файлын 3-р баримт) руу ханддаг.
servers:
  - url: http://localhost:8000/api/v1
    description: DEV (paper)
  - url: https://{host}/api/v1
    description: UAT / Prod
    variables:
      host:
        default: trader.internal
...
security:
  - operatorSession: []
paths:
...
  /account:
    get:
      operationId: getAccount
      tags: [read]
      summary: Дансны хормын хувилбар (Alpaca-ийн буцаасан утга, өөрчлөлтгүй)
      description: |
        AC-1: талбар бүр Alpaca REST-ийн хариунаас ШУУД гарна. Backend нь
        equity, buying power зэргийг локалаар ХЭЗЭЭ Ч дахин тооцохгүй.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountEnvelope"
        "503":
          $ref: "#/components/responses/BrokerUnavailable"
  /positions:
    get:
      operationId: getPositions
      tags: [read]
      summary: Нээлттэй позицууд (origin-той)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PositionsEnvelope"
        "503":
          $ref: "#/components/responses/BrokerUnavailable"
  /equity/history:
    get:
      operationId: getEquityHistory
      tags: [read]
      summary: Минут тутмын equity түүх (T-99, хувийн төсөл)
      description: |
        `snapshot_equity` job-ийн БОДИТ түүврүүд — интерполяци/тооцоолол
        БАЙХГҮЙ. `/account`-ийн raw pass-through-ийг ХЭЗЭЭ Ч орлохгүй,
        зөвхөн график зурахад зориулагдсан нэмэлт цуврал.
      parameters:
        - name: minutes
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10080
            default: 240
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquityHistoryEnvelope"
  /market/quote/{symbol}:
    get:
      operationId: getQuote
      tags: [read]
      summary: Сүүлийн quote (илгээхээс өмнөх notional-ийн эх сурвалж)
      description: |
        LLD §16.5 — гарын ticket дээрх «тооцоолсон notional» нь ЭНЭ утгаас
        бодогдоно, operator-ийн бичсэн үнээс БИШ. Quote хуучирсан бол
        `stale: true` — далдлахгүй. Quote огт байхгүй бол 503, таамагласан
        үнэ ХЭЗЭЭ Ч буцаахгүй.
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteEnvelope"
        "503":
          $ref: "#/components/responses/BrokerUnavailable"
  /market/bars/{symbol}:
    get:
      operationId: getMarketBars
      tags: [read]
      summary: Candle chart-ийн bar-ууд (T-99, хувийн төсөл)
      description: |
        Байгаа bar-уудыг л буцаана — зохиосон bar ХЭЗЭЭ Ч биш. Стock болон
        crypto pair (`BTC/USD`) хоёуланд ажиллана.
      parameters:
        - name: symbol
          in: path
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          schema:
            type: string
            default: 5Min
        - name: minutes
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 10080
            default: 240
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BarsEnvelope"
        "503":
          $ref: "#/components/responses/BrokerUnavailable"
  /orders:
    get:
      operationId: getOrders
      tags: [read, orders]
      summary: Order-ийн түүх
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, all]
            default: open
        - name: symbol
          in: query
          schema:
            type: string
        - name: origin
          in: query
          schema:
            $ref: "#/components/schemas/Origin"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrdersEnvelope"
...
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: session
      description: |
        Ганц operator-ийн session cookie (HttpOnly, Secure, SameSite=Strict).
        Олон хэрэглэгчийн role model v1-д БАЙХГҮЙ (спек A-6).
...
  responses:
...
    BrokerUnavailable:
      description: |
        Alpaca хүрэхгүй байна. Систем таамаглахгүй — хуучин утгыг live мэт
        буцаахгүй. Frontend үүнийг «stale» banner болгоно.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Money:
      type: string
      pattern: '^-?[0-9]+\.[0-9]{2}$'
      description: Fixed-point decimal тэмдэгт мөр. Float ХЭЗЭЭ Ч биш (AC-25).
      examples: ["10234.56", "-42.00"]
    Quantity:
      type: string
      pattern: '^-?[0-9]+(\.[0-9]{1,9})?$'
      description: Ширхэг. Crypto-д бутархай байж болох тул decimal тэмдэгт мөр.
    Utc:
      type: string
      format: date-time
      pattern: 'Z$'
      description: RFC 3339, ЗААВАЛ UTC (`Z`). Локал бүс хориотой (AC-26).
    Source:
      type: string
      enum: [alpaca_live, alpaca_paper, backtest]
      description: |
        Энэ хариун дахь БҮХ өгөгдлийн гарал. Нэг хариу дотор холимог утга
        БАЙХГҮЙ — холих шаардлагатай бол тусдаа хүсэлт (AC-20, AC-21).
    SystemState:
      type: string
      enum: [active, winding_down, halted]
    Origin:
      type: string
      enum: [research_agent, auto_tuning, manual_operator, external]
      description: |
        Order / позицийн гарал үүсэл (AC-29). `external` = локал `orders`-д
        таарах бичлэг ОЛДООГҮЙ (Alpaca UI-аас нээсэн г.м.). ХЭЗЭЭ Ч
        таамаглахгүй — таарахгүй бол `external`.
    RiskDecision:
      type: string
      enum: [APPROVE, REJECT, ESCALATE_TO_HUMAN]
    OrderSide:
      type: string
      enum: [buy, sell]
    OrderType:
      type: string
      enum: [market, limit, stop, stop_limit, bracket]
    TimeInForce:
      type: string
      enum: [day, gtc, ioc, fok]
    Envelope:
      type: object
      required: [source, as_of, stale, system_state]
      properties:
        source:
          $ref: "#/components/schemas/Source"
        as_of:
          $ref: "#/components/schemas/Utc"
        stale:
          type: boolean
          default: false
          description: |
            Өгөгдөл `STALE_AFTER_SECONDS`-оос хуучин. Талбарыг ОРХИХГҮЙ —
            ил `false` бичнэ (хавсралт 10 §2.4: «байхгүй» нь таамаглах цонх).
        system_state:
          $ref: "#/components/schemas/SystemState"
...
    Account:
      type: object
      required: [account_id, equity, cash, buying_power, pattern_day_trader, day_trade_count]
      properties:
        account_id:
          type: string
        equity:
          $ref: "#/components/schemas/Money"
        cash:
          $ref: "#/components/schemas/Money"
        buying_power:
          $ref: "#/components/schemas/Money"
        last_equity:
          $ref: "#/components/schemas/Money"
        pattern_day_trader:
          type: boolean
        day_trade_count:
          type: integer
        trading_blocked:
          type: boolean
    AccountEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          account:
            account_id: PA3X9QK2TLZ0
            equity: "104238.17"
            cash: "38210.55"
            buying_power: "76421.10"
            last_equity: "103980.44"
            pattern_day_trader: false
            day_trade_count: 1
            trading_blocked: false
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [account]
          properties:
            account:
              $ref: "#/components/schemas/Account"
    EquityHistoryEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-18T14:30:00Z"
          stale: false
          system_state: active
          points:
            - ts: "2026-09-18T14:28:00Z"
              equity: "100000.59"
              cash: "70000.41"
            - ts: "2026-09-18T14:29:00Z"
              equity: "100012.30"
              cash: "70000.41"
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [points]
          properties:
            points:
              type: array
              items:
                type: object
                required: [ts, equity, cash]
                properties:
                  ts:
                    $ref: "#/components/schemas/Utc"
                  equity:
                    $ref: "#/components/schemas/Money"
                  cash:
                    $ref: "#/components/schemas/Money"
...
    Quote:
      type: object
      required: [symbol, bid, ask, last, quote_ts]
      properties:
        symbol:
          type: string
        bid:
          $ref: "#/components/schemas/Money"
        ask:
          $ref: "#/components/schemas/Money"
        last:
          $ref: "#/components/schemas/Money"
        quote_ts:
          $ref: "#/components/schemas/Utc"
    BarsEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [symbol, timeframe, bars]
          properties:
            symbol:
              type: string
            timeframe:
              type: string
            bars:
              type: array
              items:
                type: object
                required: [t, open, high, low, close, volume]
                properties:
                  t:
                    $ref: "#/components/schemas/Utc"
                  open:
                    $ref: "#/components/schemas/Money"
                  high:
                    $ref: "#/components/schemas/Money"
                  low:
                    $ref: "#/components/schemas/Money"
                  close:
                    $ref: "#/components/schemas/Money"
                  volume:
                    type: integer
    QuoteEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          quote:
            symbol: AAPL
            bid: "221.48"
            ask: "221.52"
            last: "221.50"
            quote_ts: "2026-09-16T14:29:58Z"
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [quote]
          properties:
            quote:
              $ref: "#/components/schemas/Quote"
    Position:
      type: object
      required:
        [symbol, qty, side, avg_entry_price, market_value, unrealized_pl, origin, origin_mixed]
      properties:
        symbol:
          type: string
        qty:
          $ref: "#/components/schemas/Quantity"
        side:
          type: string
          enum: [long, short]
        avg_entry_price:
          $ref: "#/components/schemas/Money"
        market_value:
          $ref: "#/components/schemas/Money"
        unrealized_pl:
          $ref: "#/components/schemas/Money"
        origin:
          $ref: "#/components/schemas/Origin"
        origin_detail:
          type: [string, "null"]
          description: >-
            `research_agent` үед provider/model; `manual_operator` үед actor;
            `external` үед `null`.
        origin_mixed:
          type: boolean
          description: >-
            Энэ symbol дээр НЭГЭЭС ОЛОН origin fill хийсэн (LLD §11.1). `origin`
            нь сүүлийн fill-ийнх — бүтэн задаргаа БИШ. UI үүнийг ил тэмдэглэнэ,
            далдлахгүй (LLD §16.4).
    PositionsEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          positions:
            - symbol: AAPL
              qty: "40"
              side: long
              avg_entry_price: "221.40"
              market_value: "9012.00"
              unrealized_pl: "156.00"
              origin: research_agent
              origin_detail: claude-mcp/claude-opus-5
              origin_mixed: false
            - symbol: MSFT
              qty: "12"
              side: long
              avg_entry_price: "402.10"
              market_value: "4870.80"
              unrealized_pl: "-54.40"
              origin: manual_operator
              origin_detail: operator
              origin_mixed: true
            - symbol: TSLA
              qty: "5"
              side: long
              avg_entry_price: "248.90"
              market_value: "1252.50"
              unrealized_pl: "8.00"
              origin: external
              origin_detail: null
              origin_mixed: false
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [positions]
          properties:
            positions:
              type: array
              items:
                $ref: "#/components/schemas/Position"
    Order:
      type: object
      required: [id, client_order_id, symbol, side, qty, order_type, status, origin, submitted_at]
      properties:
        id:
          type: string
          format: uuid
          description: Локал ID
        broker_order_id:
          type: [string, "null"]
          description: Alpaca-ийн ID; илгээгдэх хүртэл `null`
        client_order_id:
          type: string
          description: Idempotency түлхүүрээс гаргасан детерминистик ID
        symbol:
          type: string
        side:
          $ref: "#/components/schemas/OrderSide"
        qty:
          $ref: "#/components/schemas/Quantity"
        filled_qty:
          $ref: "#/components/schemas/Quantity"
        order_type:
          $ref: "#/components/schemas/OrderType"
        limit_price:
          $ref: "#/components/schemas/Money"
        stop_price:
          $ref: "#/components/schemas/Money"
        time_in_force:
          $ref: "#/components/schemas/TimeInForce"
        status:
          type: string
          enum: [pending_risk, pending_approval, rejected, accepted, partially_filled, filled, canceled, expired, failed]
        origin:
          $ref: "#/components/schemas/Origin"
        origin_detail:
          type: [string, "null"]
        decision_id:
          type: [string, "null"]
          format: uuid
          description: Холбогдох `agent_decisions` мөр
        submitted_at:
          $ref: "#/components/schemas/Utc"
        filled_at:
          oneOf:
            - $ref: "#/components/schemas/Utc"
            - type: "null"
          description: Биелэх хүртэл `null`
    OrdersEnvelope:
      examples:
        - source: alpaca_paper
          as_of: "2026-09-16T14:30:00Z"
          stale: false
          system_state: active
          orders:
            - id: 6f1b5a2c-0d4e-4f31-9b77-1a2c3d4e5f60
              broker_order_id: 0b6c9a51-4e2f-4a1d-9f0c-7d8e9f0a1b2c
              client_order_id: p3-6f1b5a2c
              symbol: AAPL
              side: buy
              qty: "10"
              filled_qty: "10"
              order_type: limit
              limit_price: "221.50"
              time_in_force: day
              status: filled
              origin: research_agent
              origin_detail: claude-mcp/claude-opus-5
              decision_id: 3c9d8e7f-6a5b-4c3d-8e1f-0a9b8c7d6e5f
              submitted_at: "2026-09-16T13:55:02Z"
              filled_at: "2026-09-16T13:55:04Z"
            - id: 9a8b7c6d-5e4f-4a3b-8c2d-1e0f9a8b7c6d
              broker_order_id: null
              client_order_id: p3-9a8b7c6d
              symbol: MSFT
              side: sell
              qty: "12"
              filled_qty: "0"
              order_type: market
              time_in_force: day
              status: pending_risk
              origin: manual_operator
              origin_detail: operator
              decision_id: null
              submitted_at: "2026-09-16T14:29:51Z"
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [orders]
          properties:
            orders:
              type: array
              items:
                $ref: "#/components/schemas/Order"
...
    RiskEvaluation:
      type: object
      required: [decision, checks, evaluated_at]
      properties:
        decision:
          $ref: "#/components/schemas/RiskDecision"
        reason:
          type: [string, "null"]
        evaluated_at:
          $ref: "#/components/schemas/Utc"
        checks:
          type: array
          description: >-
            Шалгалт бүрийн үр дүн — юу унасныг, ямар хязгаартай харьцуулснаар
            нь ил харуулна.
          items:
            type: object
            required: [rule, passed]
            properties:
              rule:
                type: string
                enum:
                  - system_state
                  - restricted_symbol
                  - position_pct
                  - total_exposure_pct
                  - daily_loss_limit
                  - pdt_day_trades
                  - order_notional
                  - price_sanity
                  - wind_down_direction
              passed:
                type: boolean
              limit_name:
                type: string
                description: "config-ийн нэр, жишээ: `MAX_POSITION_PCT`"
              limit_value:
                type: string
              actual_value:
                type: string
              detail:
                type: string
...
    Problem:
      description: RFC 9457 Problem Details.
      type: object
      required: [type, title, status, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          enum:
            - confirmation_required
            - system_halted
            - winding_down_increase_blocked
            - risk_rejected
            - grounding_failed
            - idempotency_conflict
            - breaker_still_tripped
            - already_active
            - approval_expired
            - version_conflict
            - broker_unavailable
            - broker_rejected
            - provider_unavailable
            - not_found
            - invalid_request
        detail:
          type: string
        broker_code:
          type: string
          description: >-
            Зөвхөн `broker_rejected` үед. Alpaca-ийн алдааны код
            (ж: `42210000` — wash trade). Broker ХАРИУЛСАН гэдгийн нотолгоо.
        errors:
          type: array
          description: Зөвхөн `invalid_request` үед. Унасан талбар тус бүр.
          items:
            type: object
            required: [loc, msg]
            properties:
              loc:
                type: string
              msg:
                type: string
        risk:
          $ref: "#/components/schemas/RiskEvaluation"
        confirmation:
          type: object
          description: >-
            Зөвхөн `confirmation_required` үед. Хоёр шаттай баталгаажуулалтын
            сорилт.
          required: [token, prompt, expires_at]
          properties:
            token:
              type: string
            prompt:
              type: string
              description: |
                Operator-т харуулах тодорхой текст. Үйлдэл тус бүрд ӨӨР —
                хоёр товчийг андуурахаас сэргийлнэ (AC-38).
            expires_at:
              $ref: "#/components/schemas/Utc"
        breaker_metrics:
          type: array
          items:
            type: object
            additionalProperties: true